+++ cbc/key_definition_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestKeyDefinitionsWithCodes(t *testing.T) {
+func TestKeyDefinitionsWithValues(t *testing.T) {
 	kd := &cbc.KeyDefinition{
 		Key: "key",
 		Name: i18n.String{
@@ -20,49 +20,21 @@ func TestKeyDefinitionsWithCodes(t *testing.T) {
 			i18n.EN: "Description",
 			i18n.ES: "Descripción",
 		},
-		Codes: []*cbc.CodeDefinition{
+		Values: []*cbc.ValueDefinition{
 			{
-				Code: cbc.Code("CODE"),
+				Value: "value",
 				Name: i18n.String{
-					i18n.EN: "Code",
-					i18n.ES: "Código",
+					i18n.EN: "Value",
+					i18n.ES: "Valor",
 				},
 			},
 		},
 	}
-	assert.True(t, kd.HasCode(cbc.Code("CODE")))
-	assert.False(t, kd.HasCode(cbc.Code("INVALID")))
-	cd := kd.CodeDef(cbc.Code("CODE"))
-	require.NotNil(t, cd)
-	assert.Equal(t, "Code", cd.Name[i18n.EN])
-}
-
-func TestKeyDefinitionsWithKeys(t *testing.T) {
-	kd := &cbc.KeyDefinition{
-		Key: "key",
-		Name: i18n.String{
-			i18n.EN: "Name",
-			i18n.ES: "Nombre",
-		},
-		Desc: i18n.String{
-			i18n.EN: "Description",
-			i18n.ES: "Descripción",
-		},
-		Keys: []*cbc.KeyDefinition{
-			{
-				Key: cbc.Key("code"),
-				Name: i18n.String{
-					i18n.EN: "Code",
-					i18n.ES: "Código",
-				},
-			},
-		},
-	}
-	assert.True(t, kd.HasKey(cbc.Key("code")))
-	assert.False(t, kd.HasKey(cbc.Key("invalid")))
-	cd := kd.KeyDef(cbc.Key("code"))
-	require.NotNil(t, cd)
-	assert.Equal(t, "Code", cd.Name[i18n.EN])
+	assert.True(t, kd.HasValue("value"))
+	assert.False(t, kd.HasValue("invalid"))
+	vd := kd.ValueDef("value")
+	require.NotNil(t, vd)
+	assert.Equal(t, "Value", vd.Name[i18n.EN])
 }
 
 func TestKeyDefinitionWithPattern(t *testing.T) {