+++ regimes/gr/tax_combo_test.go
@@ -7,11 +7,10 @@ import (
 	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/regimes/gr"
 	"github.com/invopop/gobl/tax"
-	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestValidateTaxCombo(t *testing.T) {
-
 	greece := gr.New()
 	ctx := greece.WithContext(context.Background())
 
@@ -21,16 +20,39 @@ func TestValidateTaxCombo(t *testing.T) {
 			Percent:  num.NewPercentage(4, 2),
 		}
 		err := tc.ValidateWithContext(ctx)
-		require.ErrorContains(t, err, "ext: (gr-iapr-vat-cat: required.)")
+		assert.ErrorContains(t, err, "ext: (gr-mydata-vat-cat: required.)")
 	})
 
 	t.Run("exemption presence", func(t *testing.T) {
 		tc := &tax.Combo{
 			Category: tax.CategoryVAT,
-			Rate:     tax.RateExempt,
-			Ext:      tax.Extensions{"gr-iapr-vat-cat": "1"},
+			Ext:      tax.Extensions{"gr-mydata-vat-cat": "1"},
 		}
 		err := tc.ValidateWithContext(ctx)
-		require.ErrorContains(t, err, "gr-iapr-exemption: required")
+		assert.ErrorContains(t, err, "gr-mydata-exemption: required")
+	})
+
+	t.Run("income classification keys presence", func(t *testing.T) {
+		tx := &tax.Combo{
+			Category: tax.CategoryVAT,
+			Percent:  num.NewPercentage(24, 2),
+			Ext: tax.Extensions{
+				gr.ExtKeyMyDATAVATCat:    "1",
+				gr.ExtKeyMyDATAIncomeCat: "category1_1",
+			},
+		}
+		err := tx.ValidateWithContext(ctx)
+		assert.ErrorContains(t, err, "gr-mydata-income-type: required")
+
+		tx.Ext[gr.ExtKeyMyDATAIncomeType] = "E3_106"
+		delete(tx.Ext, gr.ExtKeyMyDATAIncomeCat)
+
+		err = tx.ValidateWithContext(ctx)
+		assert.ErrorContains(t, err, "gr-mydata-income-cat: required")
+
+		tx.Ext[gr.ExtKeyMyDATAIncomeCat] = "category1_1"
+
+		err = tx.ValidateWithContext(ctx)
+		assert.NoError(t, err)
 	})
 }